requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
openai>=1.30.0
mcp>=1.2.0
typing-extensions>=4.12.2
chromadb>=0.5.4
google-generativeai>=0.7.2


